Allocate safe regions and Implement runtime functions

+++ compiler-rt/lib/dfisan/dfisan_interface_internal.h
@@ -31,12 +31,12 @@ extern "C" {
   SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_store_id_16(uptr StoreAddr, u16 DefID);
 
   // Checks whether Def-IDs in the given range of the shadow memory are the same as correct IDs.
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_n (uptr LoadAddr, u64 Size, u16 Argc, ...);
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_1 (uptr LoadAddr, u16 Argc, ...);
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_2 (uptr LoadAddr, u16 Argc, ...);
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_4 (uptr LoadAddr, u16 Argc, ...);
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_8 (uptr LoadAddr, u16 Argc, ...);
-  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_16(uptr LoadAddr, u16 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_n (uptr LoadAddr, u64 Size, u32 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_1 (uptr LoadAddr, u32 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_2 (uptr LoadAddr, u32 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_4 (uptr LoadAddr, u32 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_8 (uptr LoadAddr, u32 Argc, ...);
+  SANITIZER_INTERFACE_ATTRIBUTE void __dfisan_check_ids_16(uptr LoadAddr, u32 Argc, ...);
 
 } // externc "C"
